Convert ffmpeg.{js,ts}
